requirements.txt
streamlit==1.33.0
faiss-cpu==1.13.2
numpy==1.26.4
sentence-transformers==2.7.0
torch==2.2.2
scikit-learn==1.4.2
ollama==0.1.8